+++ integrations/azure-devops/CHANGELOG.md
@@ -15,12 +15,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added OAuth integration support for Azure DevOps
 
 
-## 0.6.32 (2026-01-27)
+## 0.7.2 (2026-02-01)
 
 
 ### Improvements
 
-- Added OAuth configuration support for Azure DevOps integration.
+- Bumped ocean version to ^0.35.1
+
+
+## 0.7.1 (2026-02-01)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.35.0
+
+
+## 0.7.0 (2026-01-29)
+
+
+### Features
+
+- Introduced resync for `advanced-security-alert` kind
+- Implemented live events for `advanced-security-alert` kind
+
+
+## 0.6.32 (2026-01-29)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.34.0
 
 
 ## 0.6.31 (2026-01-27)